pyproject.toml
[project]
name = "product-listing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "beautifulsoup4>=4.13.4",
    "chainlit>=2.6.3",
    "dapr>=1.15.0",
    "dotenv>=0.9.9",
    "openai-agents>=0.2.3",
]
